Add language bar to layout
