+++ client/src/components/Layout/General/index.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import * as S from './style';
-
 import { useMediaQuery } from 'react-responsive';
-
 import { usePublicOrg } from '../../../context/public-org';
 import { useAdminOrg } from '../../../context/admin-org';
-
 import { OrganisationLogo } from '../../../components';
+import Language from '../../Language';
 
 // import Navbar from '../../Navbar';
 import GoBack from '../../GoBack';
@@ -22,7 +20,7 @@ const General = ({ children, goBack, maxWidth, showHelp, ...props }) => {
   });
   return (
     <S.Container>
-      accessability header goes here
+      <Language />
       <OrganisationLogo logoUrl={adminOrg?.logoUrl || publicOrg?.logoUrl} />
       {/* <S.LangButton>
             <T.P isSmall weight="bold" mr="4px">
